[RAMEN9610-10619][Common][9610] wlbt: [ACS] Rssi factor calculation issue.
Driver changes to calculate rssi factor for 2.4GHZ.

SCSC-Bug-Id: SSB-47988

+++ drivers/net/wireless/scsc/nl80211_vendor.h
@@ -1649,8 +1649,8 @@ struct slsi_acs_chan_info {
 	u8 num_bss_load_ap;
 	u8 total_chan_utilization;
 	u8 avg_chan_utilization;
-	u8 rssi_factor;
-	u8 adj_rssi_factor;
+	int rssi_factor;
+	int adj_rssi_factor;
 };
 
 struct slsi_acs_selected_channels {
